feat: setup basic plugin with example
